Add the device tree config for ADV7282, TC358743, and OV5647
These are the compiled files from
https://github.com/6by9/linux/tree/rpi-4.14.y-unicam-dt

Default configuration is set for Pi 3B and 3B+.
If using on a Pi A+, B+, 2B, 0, or 0W, then please add
"i2c_pins_28_29=1" as a parameter to the overlay.

The original model A/B is not supported due to differences in
how the camera is connected.

Load with the appropriate line from:
  dtoverlay=tc358743
  dtoverlay=adv7282m
  dtoverlay=ov5647

If you wish to capture audio from the TC358743, then also add
  dtoverlay=tc358743-audio
and connect
  SD -> Pi GPIO20 PCM_DIN
  SCK -> Pi GPIO18 PCM_CLK
  WFS -> Pi GPIO19 PCM_FS
  GND -> a Pi GND
(On the Auvidea B101 and B102 boards, those are pins 5, 6, 7,
 & 8 respectively)
The audio will then be available via ALSA, but the sample rate
must be queried from V4L2. "v4l2-ctl --list-ctrls" will list controls
"audio_sampling_rate" and "audio_present" which have the obvious
meanings.