Add example EDID files
The TC358743 needs an EDID to report to the connected
device what the supported modes are.

This commit provides two example EDID files, one for
resolutions up to 1080P50, and the other includes 1080P60
as well. 1080P60 can ONLY be achieved on a Compute Module
or Compute Module 3 as it requires at least 3 CSI2 lanes
and a normal Pi only exposes 2 lanes.